@@ -1,7 +1,7 @@
 package diecast
 
 import (
-	"io/ioutil"
+	"io"
 	"net/http"
 
 	"github.com/microcosm-cc/bluemonday"
@@ -28,7 +28,7 @@ func (self *MarkdownRenderer) SetPrewriteFunc(fn PrewriteFunc) {
 func (self *MarkdownRenderer) Render(w http.ResponseWriter, req *http.Request, options RenderOptions) error {
 	defer options.Input.Close()
 
-	if input, err := ioutil.ReadAll(options.Input); err == nil {
+	if input, err := io.ReadAll(options.Input); err == nil {
 		var output = blackfriday.Run(
 			input,
 			blackfriday.WithExtensions(blackfriday.CommonExtensions),
